+++ beacon/validator/src/main/java/tech/pegasys/teku/validator/coordinator/ValidatorApiHandler.java
@@ -543,11 +543,13 @@ public class ValidatorApiHandler implements ValidatorApiChannel {
                 LOG.trace(
                     "Successfully imported proposed block: {}",
                     () -> formatBlock(block.getSlot(), block.getRoot()));
+                dutyMetrics.onBlockPublished(block.getMessage().getSlot());
                 return SendSignedBlockResult.success(block.getRoot());
               } else if (result.getFailureReason() == FailureReason.BLOCK_IS_FROM_FUTURE) {
                 LOG.debug(
                     "Delayed processing proposed block {} because it is from the future",
                     formatBlock(block.getSlot(), block.getRoot()));
+                dutyMetrics.onBlockPublished(block.getMessage().getSlot());
                 return SendSignedBlockResult.notImported(result.getFailureReason().name());
               } else {
                 VALIDATOR_LOGGER.proposedBlockImportFailed(
@@ -662,6 +664,7 @@ public class ValidatorApiHandler implements ValidatorApiChannel {
             ? spec.atEpoch(epoch).getBeaconStateUtil().getCurrentDutyDependentRoot(state)
             : spec.atEpoch(epoch).getBeaconStateUtil().getPreviousDutyDependentRoot(state);
     return new AttesterDuties(
+        combinedChainDataClient.isChainHeadOptimistic(),
         dependentRoot,
         validatorIndices
             .intStream()