+++ ethereum/executionclient/src/integration-test/java/tech/pegasys/teku/ethereum/executionclient/rest/RestExecutionBuilderClientTest.java
@@ -144,7 +144,7 @@ class RestExecutionBuilderClientTest {
   }
 
   @TestTemplate
-  void registerValidator_success() {
+  void registerValidators_success() {
 
     mockWebServer.enqueue(new MockResponse().setResponseCode(200));
 
@@ -163,7 +163,24 @@ class RestExecutionBuilderClientTest {
   }
 
   @TestTemplate
-  void registerValidator_failures() {
+  void registerValidators_zeroRegistrationsDoesNotMakeRequest() {
+
+    SszList<SignedValidatorRegistration> zeroRegistrations =
+        SIGNED_VALIDATOR_REGISTRATIONS_SCHEMA.getDefault();
+
+    assertThat(restExecutionBuilderClient.registerValidators(SLOT, zeroRegistrations))
+        .succeedsWithin(WAIT_FOR_CALL_COMPLETION)
+        .satisfies(
+            response -> {
+              assertThat(response.isSuccess()).isTrue();
+              assertThat(response.getPayload()).isNull();
+            });
+
+    assertThat(mockWebServer.getRequestCount()).isEqualTo(0);
+  }
+
+  @TestTemplate
+  void registerValidators_failures() {
 
     String unknownValidatorError = "{\"code\":400,\"message\":\"unknown validator\"}";