+++ ethereum/networks/src/main/java/tech/pegasys/teku/networks/Eth2NetworkConfiguration.java
@@ -43,7 +43,7 @@ public class Eth2NetworkConfiguration {
   private static final int DEFAULT_STARTUP_TARGET_PEER_COUNT = 5;
   private static final int DEFAULT_STARTUP_TIMEOUT_SECONDS = 30;
   public static final boolean DEFAULT_PROPOSER_BOOST_ENABLED = true;
-  public static final boolean DEFAULT_EQUIVOCATING_INDICES_ENABLED = false;
+  public static final boolean DEFAULT_EQUIVOCATING_INDICES_ENABLED = true;
   public static final boolean DEFAULT_FORK_CHOICE_BEFORE_PROPOSING_ENABLED = false;
 
   private final Spec spec;
@@ -419,20 +419,25 @@ public class Eth2NetworkConfiguration {
       discoveryBootnodes = new ArrayList<>();
       eth1DepositContractAddress = null;
       eth1DepositContractDeployBlock = Optional.empty();
+      forkChoiceBeforeProposingEnabled = DEFAULT_FORK_CHOICE_BEFORE_PROPOSING_ENABLED;
 
       return this;
     }
 
+    public Builder applyTestnetDefaults() {
+      return reset().forkChoiceBeforeProposingEnabled(true);
+    }
+
     public Builder applyMinimalNetworkDefaults() {
-      return reset().constants(MINIMAL.configName()).startupTargetPeerCount(0);
+      return applyTestnetDefaults().constants(MINIMAL.configName()).startupTargetPeerCount(0);
     }
 
     public Builder applySwiftNetworkDefaults() {
-      return reset().constants(SWIFT.configName()).startupTargetPeerCount(0);
+      return applyTestnetDefaults().constants(SWIFT.configName()).startupTargetPeerCount(0);
     }
 
     public Builder applyLessSwiftNetworkDefaults() {
-      return reset().constants(LESS_SWIFT.configName()).startupTargetPeerCount(0);
+      return applyTestnetDefaults().constants(LESS_SWIFT.configName()).startupTargetPeerCount(0);
     }
 
     public Builder applyMainnetNetworkDefaults() {
@@ -465,7 +470,7 @@ public class Eth2NetworkConfiguration {
     }
 
     public Builder applyPraterNetworkDefaults() {
-      return reset()
+      return applyTestnetDefaults()
           .constants(PRATER.configName())
           .startupTimeoutSeconds(120)
           .eth1DepositContractDeployBlock(4367322)
@@ -488,7 +493,7 @@ public class Eth2NetworkConfiguration {
     }
 
     public Builder applyRopstenNetworkDefaults() {
-      return reset()
+      return applyTestnetDefaults()
           .constants(ROPSTEN.configName())
           .startupTimeoutSeconds(120)
           .eth1DepositContractDeployBlock(12269949)
@@ -502,7 +507,8 @@ public class Eth2NetworkConfiguration {
     }
 
     public Builder applyKilnNetworkDefaults() {
-      return reset()
+      return applyTestnetDefaults()
+          .forkChoiceBeforeProposingEnabled(true)
           .constants(KILN.configName())
           .startupTimeoutSeconds(120)
           .eth1DepositContractDeployBlock(0)