+++ CHANGELOG.md
@@ -17,6 +17,9 @@ For information on changes in released versions of Teku, see the [releases page]
 
 ### Additions and Improvements
 - Reduced memory requirements for storing the deposit merkle tree.
+- Enable spec change to ignore weightings from attestations from equivocating validators.
+- Fork choice before proposals is now enabled by default for testnets. It can be disabled with `--Xfork-choice-before-proposing-enabled=false` if required.
 
 ### Bug Fixes
 - Fixed issue where the REST API may return content as SSZ instead of JSON if the header `Accept: */*` was specified.
+- Fixed issue where sync committee aggregations were skipped, but reported failed because there were no signatures to aggregate.